Добавил тесты на util для int8, прописал несколько беньчмарков, исправил типы в utils

+++ implementation/src/main/java/org/column4j/utils/Float64VectorUtils.java
@@ -23,10 +23,10 @@ public final class Float64VectorUtils {
      * @param tombstone ignored value
      * @param from      from index (inclusive)
      * @param to        to index (exclusive)
-     * @return maximum on array or {@link Float#MIN_VALUE}
+     * @return maximum on array or {@link Double.MIN_VALUE}
      */
     public static double max(double[] data, double tombstone, int from, int to) {
-        var maxVector = DoubleVector.broadcast(SPECIES_PREFERRED, Float.MIN_VALUE);
+        var maxVector = DoubleVector.broadcast(SPECIES_PREFERRED, Double.MIN_VALUE);
         for (; from < to && from + SPECIES_LENGTH <= to; from += SPECIES_LENGTH) {
             var nextVector = DoubleVector.fromArray(SPECIES_PREFERRED, data, from);
             var tombstoneMask = nextVector.eq(tombstone).not();
@@ -48,10 +48,10 @@ public final class Float64VectorUtils {
      * @param tombstone ignored value
      * @param from      from index (inclusive)
      * @param to        to index (exclusive)
-     * @return minimum on array or {@link Float#MAX_VALUE}
+     * @return minimum on array or {@link Double.MAX_VALUE}
      */
     public static double min(double[] data, double tombstone, int from, int to) {
-        var minVector = DoubleVector.broadcast(SPECIES_PREFERRED, Float.MAX_VALUE);
+        var minVector = DoubleVector.broadcast(SPECIES_PREFERRED, Double.MAX_VALUE);
         for (; from < to && from + SPECIES_LENGTH <= to; from += SPECIES_LENGTH) {
             var nextVector = DoubleVector.fromArray(SPECIES_PREFERRED, data, from);
             var tombstoneMask = nextVector.eq(tombstone).not();
